+++ src/serde/__init__.py
@@ -55,17 +55,18 @@ found in `~serde.field`.
 
 from .error import DeserializationError, SerdeError, SerializationError, ValidationError
 from .field import (
-    Bool, Boolean, Choice, Date, DateTime, Dict, Dictionary, Domain, Email, Field, Float,
-    Instance, Int, Integer, List, Nested, Slug, Str, String, Time, Tuple, Url, Uuid
+    Bool, Boolean, Choice, Date, DateTime, Dict, Dictionary, Domain, Email,
+    Field, Float, Instance, Int, Integer, IpAddress, Ipv4Address, Ipv6Address,
+    List, MacAddress, Nested, Slug, Str, String, Time, Tuple, Url, Uuid
 )
 from .model import Model
 
 
 __all__ = [
     'Bool', 'Boolean', 'Choice', 'Date', 'DateTime', 'DeserializationError', 'Dict', 'Dictionary',
-    'Domain', 'Email', 'Field', 'Float', 'Instance', 'Int', 'Integer', 'List', 'Model', 'Nested',
-    'SerdeError', 'SerializationError', 'Slug', 'Str', 'String', 'Time', 'Tuple', 'Url', 'Uuid',
-    'ValidationError'
+    'Domain', 'Email', 'Field', 'Float', 'Instance', 'Int', 'Integer', 'IpAddress', 'Ipv4Address',
+    'Ipv6Address', 'List', 'MacAddress', 'Model', 'Nested', 'SerdeError', 'SerializationError',
+    'Slug', 'Str', 'String', 'Time', 'Tuple', 'Url', 'Uuid', 'ValidationError'
 ]
 __title__ = 'serde'
 __version__ = '0.2.0'